feat: converge authorization and runtime seams, add opt-in core profile
- Auth: direct grant/revoke delta via new AuthorizationMutationService and
  RoleMutationService with expectedVersion CAS, audit outbox, and receipts;
  full-replace ChangeAuthorizationCommand and UserRole write path removed
- App: app-only ports localized out of app-api into private module seams;
  /run routes through InteractiveCodeRunner; judge runtime adds bounded
  async submit/poll/cancel with optional Judge0 adapter (default off)
- Core: opt-in services/core parent with five-owner datasource wiring,
  CAS ownership handoff, and readiness endpoint; platform/observability
  OTLP HTTPS guard; search worker SearchIndex seam adapter
- Ops: managed data-layer and Redis TLS env entries, observability-managed
  compose overlay, .devcontainer, and ulticode launcher script
- Docs: ADR-0009/0010, current-status, runbook, and services issues

+++ services/app/app-web/src/main/java/com/ulticode/modules/follow/port/FollowCountPort.java
@@ -27,12 +27,12 @@ public interface FollowCountPort {
      * @return number of users following {@code userId}, or 0 if user
      *         does not exist or has no followers
      */
-    int countFollowers(String userId);
+    long countFollowers(String userId);
 
     /**
      * @param userId the user whose following count is requested
      * @return number of users {@code userId} follows, or 0 if user
      *         does not exist or follows no one
      */
-    int countFollowing(String userId);
+    long countFollowing(String userId);
 }
